feat: Add secret DIY dice roll override
Introduces a hidden "DIY Mode" that allows players to manually select the dice roll outcome after a secret sequence of clicks on the Ludo board. This feature provides a fun, hidden easter egg and adds a layer of interactivity for persistent players.

The DIY mode is activated via a click sequence on the Ludo board. Once unlocked, it intercepts dice rolls, presenting a UI to select a number between 1 and 6. This selection replaces the random dice roll.